Add 'install' command to Justfile
The 'install' command has been added to the Justfile. This command builds and installs the binary using 'cargo install' command. It is now easier for users to quickly install the binary and start using it without having to manually build it from source.

No testing was done for this commit.

Note that the other tasks in the Justfile, 'format' and 'clippy', can now be run using the 'fix' command, as described in the file.